+++ src/specex_hermite.cc
@@ -1,5 +1,5 @@
 
-#include <unbls.h>
+#include <specex_unbls.h>
 
 #include <specex_hermite.h>
