+++ src/specex_hermite.h
@@ -1,7 +1,7 @@
 #ifndef SPECEX_HERMITE__H
 #define SPECEX_HERMITE__H
 
-#include <unbls.h>
+#include <specex_unbls.h>
 
 namespace specex {
   double HermitePol(const int Degree, const double &x);
