+++ src/specex_legendre.h
@@ -3,7 +3,7 @@
 
 #include <string>
 
-#include <unbls.h>
+#include <specex_unbls.h>
 
 using namespace std;
