+++ src/specex_mask.cc
@@ -1,5 +1,5 @@
 
-#include <unbls.h>
+#include <specex_unbls.h>
 
 #include <specex_mask.h>
 #include <specex_psf.h>
