+++ src/specex_psfpy.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include <string>
 
-#include <unbls.h>
+#include <specex_unbls.h>
 #include "specex_fits.h"
 #include "specex_image_data.h"
