+++ src/specex_pyio.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include <string>
 
-#include <unbls.h>
+#include <specex_unbls.h>
 
 #include <specex_pyoptions.h>
 #include <specex_pyimage.h>
