+++ src/specex_pyprior.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <string>
 
-#include <unbls.h>
+#include <specex_unbls.h>
 
 #include <specex_pyoptions.h>
 #include <specex_psf.h>
