+++ src/specex_pypsf.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include <string>
 
-#include <unbls.h>
+#include <specex_unbls.h>
 
 #include <specex_psf.h>
 #include <specex_spot.h>
